+++ fixtures/remix-1.19.3/epic-stack/metronome.config.ts
@@ -1,12 +1,12 @@
-import { type MetronomeConfig } from "@metronome/config";
+import { type MetronomeConfig } from '@metronome-sh/config'
 
 export default {
-  ignoredRoutes: [],
-  ignoredPathnames: ["/healthcheck"],
-  ignoreHeadMethod: true,
-  // @ts-ignore
-  doNotTrack: (request) => {
-    // Track all events
-    return false;
-  },
-} satisfies MetronomeConfig;
+	ignoredRoutes: [],
+	ignoredPathnames: ['/healthcheck'],
+	ignoreHeadMethod: true,
+	// @ts-ignore
+	doNotTrack: request => {
+		// Track all events
+		return false
+	},
+} satisfies MetronomeConfig