+++ packages/config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @metronome-sh/config
 
+## 7.1.0-next.6
+
+### Patch Changes
+
+- Updated ts template
+
 ## 7.1.0-next.5
 
 ### Patch Changes