+++ packages/express/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @metronome-sh/express
 
+## 7.1.0-next.6
+
+### Patch Changes
+
+- Updated ts template
+- Updated dependencies
+  - @metronome-sh/runtime@7.1.0-next.6
+  - @metronome-sh/config@7.1.0-next.6
+  - @metronome-sh/node@7.1.0-next.6
+  - @metronome-sh/cli@7.1.0-next.6
+
 ## 7.1.0-next.5
 
 ### Patch Changes